+++ runescape/src/main/java/Class169.java
@@ -36,7 +36,7 @@ public final class Class169 {
 
     @OriginalMember(owner = "client!hm", name = "a", descriptor = "(BI)Lclient!rq;")
     public Class93_Sub2 method3529(@OriginalArg(1) int arg0) {
-        @Pc(10) Object local10 = this.aReferenceCache_85.get((long) arg0);
+        @Pc(10) Object local10 = this.aReferenceCache_85.get(arg0);
         if (local10 != null) {
             return (Class93_Sub2) local10;
         } else if (this.anTextureSource_4.textureAvailable(arg0)) {
@@ -56,7 +56,7 @@ public final class Class169 {
                 local90 = new Class93_Sub2(this.aClass19_Sub3_17, 3553, 6408, local52, local52, local33.aByte53 != 0, local114, 0, 0, false);
             }
             local90.method2946(local33.aBoolean236, local33.aBoolean235);
-            this.aReferenceCache_85.put(local90, (long) arg0);
+            this.aReferenceCache_85.put(local90, arg0);
             return local90;
         } else {
             return null;
